蓝盾流水线自动打准备发布灰度tag - gray/v1.1.13.6-2020-09-24

+++ src/backend/ci/ext/tencent/artifactory/biz-artifactory-tencent/src/main/kotlin/com/tencent/devops/artifactory/service/bkrepo/BkRepoCustomDirService.kt
@@ -31,14 +31,12 @@ import com.tencent.devops.artifactory.pojo.FileDetail
 import com.tencent.devops.artifactory.pojo.FileInfo
 import com.tencent.devops.artifactory.pojo.PathList
 import com.tencent.devops.artifactory.service.CustomDirService
+import com.tencent.devops.artifactory.service.PipelineService
 import com.tencent.devops.artifactory.util.JFrogUtil
 import com.tencent.devops.artifactory.util.PathUtils
 import com.tencent.devops.artifactory.util.RepoUtils
 import com.tencent.devops.common.api.exception.OperationException
-import com.tencent.devops.common.api.exception.PermissionForbiddenException
 import com.tencent.devops.common.archive.client.BkRepoClient
-import com.tencent.devops.common.auth.api.BSAuthProjectApi
-import com.tencent.devops.common.auth.code.BSRepoAuthServiceCode
 import org.glassfish.jersey.media.multipart.FormDataContentDisposition
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
@@ -49,12 +47,12 @@ import javax.ws.rs.NotFoundException
 
 @Service
 class BkRepoCustomDirService @Autowired constructor(
-    private val authProjectApi: BSAuthProjectApi,
-    private val bkRepoClient: BkRepoClient,
-    private val artifactoryAuthServiceCode: BSRepoAuthServiceCode
+    private val pipelineService: PipelineService,
+    private val bkRepoClient: BkRepoClient
 ) : CustomDirService {
     override fun list(userId: String, projectId: String, argPath: String): List<FileInfo> {
         logger.info("list, userId: $userId, projectId: $projectId, argPath: $argPath")
+        pipelineService.validatePermission(userId, projectId)
         val normalizedPath = PathUtils.checkAndNormalizeAbsPath(argPath)
         val fileList = bkRepoClient.listFile(
             userId,
@@ -71,6 +69,7 @@ class BkRepoCustomDirService @Autowired constructor(
 
     override fun show(userId: String, projectId: String, argPath: String): FileDetail {
         logger.info("show, userId: $userId, projectId: $projectId, argPath: $argPath")
+        pipelineService.validatePermission(userId, projectId)
         val normalizedPath = PathUtils.checkAndNormalizeAbsPath(argPath)
         val fileDetail = bkRepoClient.getFileDetail(
             userId,
@@ -81,19 +80,22 @@ class BkRepoCustomDirService @Autowired constructor(
         return RepoUtils.toFileDetail(fileDetail)
     }
 
-    override fun deploy(userId: String, projectId: String, argPath: String, inputStream: InputStream, disposition: FormDataContentDisposition) {
-        logger.info("deploy file, userId: $userId, projectId: $projectId, path: $argPath")
-        bkRepoClient.uploadFile(userId, projectId, RepoUtils.CUSTOM_REPO, argPath, inputStream)
+    override fun deploy(userId: String, projectId: String, argPath: String, inputStream: InputStream, disposition: FormDataContentDisposition, fileSizeLimitInMB: Int) {
+        logger.info("deploy file, userId: $userId, projectId: $projectId, path: $argPath, originFileName: ${disposition.fileName}, fileSizeLimitInMB: $fileSizeLimitInMB")
+        pipelineService.validatePermission(userId, projectId)
+        bkRepoClient.uploadFile(userId, projectId, RepoUtils.CUSTOM_REPO, argPath, inputStream, fileSizeLimitInMB = fileSizeLimitInMB)
     }
 
     override fun mkdir(userId: String, projectId: String, argPath: String) {
         logger.info("mkdir, userId: $userId, projectId: $projectId, argPath: $argPath")
         val normalizedPath = PathUtils.checkAndNormalizeAbsPath(argPath)
+        pipelineService.validatePermission(userId, projectId)
         bkRepoClient.mkdir(userId, projectId, RepoUtils.CUSTOM_REPO, normalizedPath)
     }
 
     override fun rename(userId: String, projectId: String, fromPath: String, toPath: String) {
         logger.info("rename, userId: $userId, projectId: $projectId, srcPath: $fromPath, toPath: $toPath")
+        pipelineService.validatePermission(userId, projectId)
         val normalizedFromPath = PathUtils.checkAndNormalizeAbsPath(fromPath)
         val normalizedToPath = PathUtils.checkAndNormalizeAbsPath(toPath)
         bkRepoClient.rename(userId, projectId, RepoUtils.CUSTOM_REPO, normalizedFromPath, normalizedToPath)
@@ -101,6 +103,7 @@ class BkRepoCustomDirService @Autowired constructor(
 
     override fun copy(userId: String, projectId: String, combinationPath: CombinationPath) {
         logger.info("copy, userId: $userId, projectId: $projectId, combinationPath: $combinationPath")
+        pipelineService.validatePermission(userId, projectId)
         val normalizeDestPath = PathUtils.checkAndNormalizeAbsPath(combinationPath.destPath)
 
         if (combinationPath.srcPaths.size > 1) {
@@ -131,13 +134,12 @@ class BkRepoCustomDirService @Autowired constructor(
 
     override fun move(userId: String, projectId: String, combinationPath: CombinationPath) {
         logger.info("move, projectId: $projectId, combinationPath: $combinationPath")
+        pipelineService.validatePermission(userId, projectId)
         val normalizedDestPath = PathUtils.checkAndNormalizeAbsPath(combinationPath.destPath)
-
         combinationPath.srcPaths.map { srcPath ->
             val normalizedSrcPath = JFrogUtil.normalize(srcPath)
 
-            if (normalizedSrcPath == normalizedDestPath ||
-                JFrogUtil.getParentFolder(normalizedSrcPath) == normalizedDestPath) {
+            if (normalizedSrcPath == normalizedDestPath || JFrogUtil.getParentFolder(normalizedSrcPath) == normalizedDestPath) {
                 logger.error("Cannot move in same path ($normalizedSrcPath, $normalizedDestPath)")
                 throw BadRequestException("不能移动到当前目录")
             }
@@ -159,6 +161,7 @@ class BkRepoCustomDirService @Autowired constructor(
 
     override fun delete(userId: String, projectId: String, pathList: PathList) {
         logger.info("delete, projectId: $projectId, pathList: $pathList")
+        pipelineService.validatePermission(userId, projectId)
         pathList.paths.map { path ->
             val normalizedPath = PathUtils.checkAndNormalizeAbsPath(path)
             bkRepoClient.delete(
@@ -168,16 +171,6 @@ class BkRepoCustomDirService @Autowired constructor(
                 normalizedPath
             )
         }
-    }
-
-    override fun validatePermission(userId: String, projectId: String) {
-        if (!isProjectUser(userId, projectId)) {
-            throw PermissionForbiddenException("用户($userId)不是工程($projectId)成员")
-        }
-    }
-
-    override fun isProjectUser(user: String, projectId: String): Boolean {
-        return authProjectApi.getProjectUsers(artifactoryAuthServiceCode, projectId).contains(user)
     }
 
     companion object {
